requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
cffi==1.10.0
chardet==3.0.4
discord.py==0.16.9
multidict==3.1.3
pycparser==2.18
six==1.10.0
websockets==3.3
yarl==0.12.0